Potential Final Folder
Getting a good level one started. Assets are garbage, but there's
gameplay. Character moves and collides with walls. Light flashes when
you start the map to display it. Door remains rigid until you hit the
switch which removes collision and changes the sprite. Exit transitions
you to a different scene (the same, in this case)